move utility from sat to utils bundle
