+++ bundles/dev.arcovia.mitigation.ilp/src/dev/arcovia/mitigation/ilp/OptimizationManager.java
@@ -29,7 +29,7 @@ import org.dataflowanalysis.dfd.dataflowdiagram.dataflowdiagramFactory;
 import dev.arcovia.mitigation.sat.CompositeLabel;
 import dev.arcovia.mitigation.sat.LabelCategory;
 import dev.arcovia.mitigation.sat.NodeLabel;
-import dev.arcovia.mitigation.sat.timeMeasurement;
+import dev.arcovia.mitigation.utils.core.timeMeasurement;
 
 public class OptimizationManager {
 	private final DataFlowDiagramAndDictionary dfd;
