+++ bundles/dev.arcovia.mitigation.utils/src/dev/arcovia/mitigation/utils/core/RunConfig.java
@@ -1,4 +1,4 @@
-package dev.arcovia.mitigation.sat;
+package dev.arcovia.mitigation.utils.core;
 
 
 public record RunConfig(
