+++ bundles/dev.arcovia.mitigation.utils/src/dev/arcovia/mitigation/utils/core/Scaler.java
@@ -1,4 +1,4 @@
-package dev.arcovia.mitigation.sat;
+package dev.arcovia.mitigation.utils.core;
 
 import java.util.HashMap;
 import java.util.HashSet;
