+++ tests/dev.arcovia.mitigation.sat.tests/META-INF/MANIFEST.MF
@@ -3,6 +3,7 @@ Bundle-ManifestVersion: 2
 Bundle-Name: SatTests
 Bundle-SymbolicName: dev.arcovia.mitigation.sat.tests
 Bundle-Version: 1.0.0.qualifier
+Export-Package: dev.arcovia.mitigation.sat.tests
 Import-Package: org.junit.jupiter.api;version="5.8.1",
  org.junit.jupiter.api.function;version="5.8.1",
  org.junit.jupiter.params;version="5.8.1",
