+++ tests/dev.arcovia.mitigation.sat.tests/testresults/ICSE.cnf
@@ -1,4 +1,4 @@
-p cnf 113 166
+p cnf 108 161
 -1 -2 3 0
 -4 5 0
 -6 -7 8 0
@@ -20,148 +20,143 @@ p cnf 113 166
 32 0
 33 0
 34 0
-35 0
+-35 0
 -36 0
 -37 0
--38 0
-39 0
+38 0
+-39 0
 -40 0
 -41 0
--42 0
-43 0
--44 0
-45 0
+42 0
+-43 0
+44 0
+-45 0
 -46 0
 -47 0
 -48 0
--49 0
-50 0
--51 0
+49 0
+-50 0
+51 0
 -52 0
 -53 0
 -54 0
--55 0
-56 0
--57 0
--58 0
--59 0
--61 -39 60 0
--60 61 0
--60 39 0
--27 -39 20 0
+-56 -38 55 0
+-55 56 0
+-55 38 0
+-27 -38 20 0
 -20 27 0
--20 39 0
--63 -39 62 0
--62 63 0
--62 39 0
--28 -39 19 0
+-20 38 0
+-58 -38 57 0
+-57 58 0
+-57 38 0
+-28 -38 19 0
 -19 28 0
--19 39 0
--65 -39 64 0
+-19 38 0
+-60 -38 59 0
+-59 60 0
+-59 38 0
+-62 -42 61 0
+-61 62 0
+-61 42 0
+-63 -42 18 0
+-18 63 0
+-18 42 0
+-65 -42 64 0
 -64 65 0
--64 39 0
--67 -43 66 0
--66 67 0
--66 43 0
--68 -43 18 0
--18 68 0
--18 43 0
--70 -43 69 0
--69 70 0
--69 43 0
--29 -43 16 0
+-64 42 0
+-29 -42 16 0
 -16 29 0
--16 43 0
--72 -43 71 0
+-16 42 0
+-67 -42 66 0
+-66 67 0
+-66 42 0
+-69 -44 68 0
+-68 69 0
+-68 44 0
+-70 -44 8 0
+-8 70 0
+-8 44 0
+-72 -44 71 0
 -71 72 0
--71 43 0
--74 -45 73 0
--73 74 0
--73 45 0
--75 -45 8 0
--8 75 0
--8 45 0
--77 -45 76 0
--76 77 0
--76 45 0
--30 -45 6 0
+-71 44 0
+-30 -44 6 0
 -6 30 0
--6 45 0
--79 -45 78 0
+-6 44 0
+-74 -44 73 0
+-73 74 0
+-73 44 0
+-76 -49 75 0
+-75 76 0
+-75 49 0
+-77 -49 13 0
+-13 77 0
+-13 49 0
+-79 -49 78 0
 -78 79 0
--78 45 0
--81 -50 80 0
--80 81 0
--80 50 0
--82 -50 13 0
--13 82 0
--13 50 0
--84 -50 83 0
--83 84 0
--83 50 0
--32 -50 11 0
+-78 49 0
+-32 -49 11 0
 -11 32 0
--11 50 0
--86 -50 85 0
+-11 49 0
+-81 -49 80 0
+-80 81 0
+-80 49 0
+-83 -51 82 0
+-82 83 0
+-82 51 0
+-84 -51 3 0
+-3 84 0
+-3 51 0
+-86 -51 85 0
 -85 86 0
--85 50 0
--88 -56 87 0
+-85 51 0
+-34 -51 1 0
+-1 34 0
+-1 51 0
+-88 -51 87 0
 -87 88 0
--87 56 0
--89 -56 3 0
+-87 51 0
+-82 97 0
+-97 82 0
+-68 98 0
+-98 68 0
+-75 99 0
+-99 75 0
+-61 100 0
+-55 100 0
+-100 61 55 0
 -3 89 0
--3 56 0
--91 -56 90 0
--90 91 0
--90 56 0
--35 -56 1 0
--1 35 0
--1 56 0
--93 -56 92 0
--92 93 0
--92 56 0
--87 102 0
--102 87 0
--73 103 0
--103 73 0
--80 104 0
--104 80 0
--66 105 0
--60 105 0
--105 66 60 0
--3 94 0
--94 3 0
--8 95 0
--95 8 0
--13 96 0
--96 13 0
--18 97 0
--20 97 0
--97 18 20 0
--90 98 0
--98 90 0
--76 99 0
--99 76 0
--83 100 0
--100 83 0
--69 101 0
--62 101 0
--101 69 62 0
--1 106 0
--106 1 0
--6 107 0
--107 6 0
--11 108 0
--108 11 0
--16 109 0
--19 109 0
--109 16 19 0
--92 110 0
--110 92 0
--78 111 0
--111 78 0
--85 112 0
--112 85 0
--71 113 0
--64 113 0
--113 71 64 0
+-89 3 0
+-8 90 0
+-90 8 0
+-13 91 0
+-91 13 0
+-18 92 0
+-20 92 0
+-92 18 20 0
+-85 93 0
+-93 85 0
+-71 94 0
+-94 71 0
+-78 95 0
+-95 78 0
+-64 96 0
+-57 96 0
+-96 64 57 0
+-1 101 0
+-101 1 0
+-6 102 0
+-102 6 0
+-11 103 0
+-103 11 0
+-16 104 0
+-19 104 0
+-104 16 19 0
+-87 105 0
+-105 87 0
+-73 106 0
+-106 73 0
+-80 107 0
+-107 80 0
+-66 108 0
+-59 108 0
+-108 66 59 0
