+++ tests/dev.arcovia.mitigation.utils.tests/META-INF/MANIFEST.MF
@@ -1,0 +1,22 @@
+Manifest-Version: 1.0
+Bundle-ManifestVersion: 2
+Bundle-Name: Tests
+Bundle-SymbolicName: dev.arcovia.mitigation.utils.tests
+Bundle-Version: 1.0.0.qualifier
+Import-Package: org.junit.jupiter.api;version="5.8.1",
+ org.junit.jupiter.api.function;version="5.8.1",
+ org.junit.jupiter.params;version="5.8.1",
+ org.junit.jupiter.params.provider;version="5.8.1"
+Require-Bundle: dev.arcovia.mitigation.sat,
+ dev.arcovia.mitigation.utils,
+ org.apache.log4j,
+ tools.mdsd.library.standalone.initialization,
+ org.dataflowanalysis.dfd.datadictionary,
+ org.dataflowanalysis.dfd.dataflowdiagram,
+ org.dataflowanalysis.analysis,
+ org.dataflowanalysis.examplemodels,
+ org.dataflowanalysis.converter,
+ org.dataflowanalysis.dfd.datadictionary,
+ dev.arcovia.mitigation.sat.tests
+Automatic-Module-Name: dev.arcovia.mitigation.utils.tests
+Bundle-RequiredExecutionEnvironment: JavaSE-17
