+++ tests/dev.arcovia.mitigation.utils.tests/src/dev/arcovia/mitigation/utils/dsl/tests/StructureTest.java
@@ -1,9 +1,9 @@
-package dev.arcovia.mitigation.sat.dsl.tests;
+package dev.arcovia.mitigation.utils.dsl.tests;
 
 import dev.arcovia.mitigation.utils.dsl.CNFTranslation;
-import dev.arcovia.mitigation.sat.dsl.tests.utility.DataLoader;
-import dev.arcovia.mitigation.sat.dsl.tests.utility.ReadabilityTestResult;
-import dev.arcovia.mitigation.sat.dsl.tests.utility.StructureResult;
+import dev.arcovia.mitigation.utils.dsl.tests.utility.DataLoader;
+import dev.arcovia.mitigation.utils.dsl.tests.utility.ReadabilityTestResult;
+import dev.arcovia.mitigation.utils.dsl.tests.utility.StructureResult;
 
 import org.apache.log4j.Logger;
 import org.dataflowanalysis.analysis.dsl.AnalysisConstraint;
