+++ tests/dev.arcovia.mitigation.utils.tests/src/dev/arcovia/mitigation/utils/dsl/tests/utility/StructureResult.java
@@ -1,4 +1,4 @@
-package dev.arcovia.mitigation.sat.dsl.tests.utility;
+package dev.arcovia.mitigation.utils.dsl.tests.utility;
 
 public record StructureResult(int dataPos, int dataNeg, int nodePos, int nodeNeg, int inputLiterals, int outputClauses, int outputLiterals, int outputLongestClause,
         int outputLiteralsPerClause, int runtime) {
